LPS-68601 Moves code to new class
